dom/tests/mochitest/whatwg/test_postMessage_override.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=postMessage
-->
<head>
  <title>postMessage override test</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=postMessage">Mozilla Bug 387706</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<iframe src="http://example.org:8000/tests/dom/tests/mochitest/whatwg/postMessage_override_helper.html"></iframe>


<pre id="test">
<script class="testbody" type="application/javascript">
/** Test for Bug 387706 **/

SimpleTest.waitForExplicitFinish();

var count = 0;

function test()
{
  try
  {
    count++;
    switch (count)
    {
      case 1:
        window.frames[0].postMessage("PASS 1");
        break;

      case 2:
        window.frames[0].postMessage("PASS 2");
        break;
  
      default:
        ok(false, "unexpected");
    }
  }
  catch (e)
  {
    ok(false, "error running test " + count + ": " + e);
  }
}

function messageReceiver(evt)
{
  var expect;

  switch (count)
  {
    case 1:
      expect = "PASS 1";
      break;

    case 2:
      expect = "PASS 2";
      break;

    default:
      expect = "unexpected";
      break;
  }

  is(evt.data, expect, "uh-oh, we didn't get the right postMessage!");
}

document.addEventListener("message", messageReceiver, false);

addLoadEvent(test);
addLoadEvent(test);
addLoadEvent(SimpleTest.finish);
</script>
</pre>
</body>
</html>
